% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{createSummarizedExperiment}
\alias{createSummarizedExperiment}
\title{Function to read the RNAseq and Meta data and create a SummarizedExperiment
Object}
\usage{
createSummarizedExperiment(pathToAssayData, pathToColData)
}
\arguments{
\item{pathToAssayData}{path to RNAseq data}

\item{pathToColData}{path to meta data file}
}
\description{
Function to read the RNAseq and Meta data and create a SummarizedExperiment
Object
}
